refactor: modularize integration source selection
