+++ src/apm_cli/integration/skill_integrator.py
@@ -14,12 +14,49 @@ from typing import TYPE_CHECKING
 from apm_cli.core.deployment_state import MaterializationResult
 from apm_cli.install.services import enforce_agent_plugin_deployment_boundary
 from apm_cli.integration.base_integrator import BaseIntegrator
+from apm_cli.integration.skill_package_routing import (
+    get_effective_type,
+)
+from apm_cli.integration.skill_package_routing import (
+    should_compile_instructions as _should_compile_instructions,
+)
+from apm_cli.integration.skill_package_routing import (
+    should_install_skill as _should_install_skill,
+)
 from apm_cli.integration.targets import TargetProfile
 from apm_cli.models.dependency.subsets import skill_subset_filter_tokens
 from apm_cli.utils.atomic_io import write_text_lf
 
+__all__ = [
+    "SkillIntegrationResult",
+    "SkillIntegrator",
+    "copy_skill_to_target",
+    "get_effective_type",
+    "normalize_skill_name",
+    "should_compile_instructions",
+    "should_install_skill",
+    "to_hyphen_case",
+    "validate_skill_name",
+]
+
 if TYPE_CHECKING:
     from apm_cli.install.deployable_source_plan import DeployableSourcePlan
+
+
+def should_install_skill(package_info) -> bool:
+    """Return whether a package should be installed as a native skill."""
+    return _should_install_skill(
+        package_info,
+        resolve_effective_type=get_effective_type,
+    )
+
+
+def should_compile_instructions(package_info) -> bool:
+    """Return whether package instructions belong in compiled output."""
+    return _should_compile_instructions(
+        package_info,
+        resolve_effective_type=get_effective_type,
+    )
 
 
 def _build_copy_ignore(
@@ -210,113 +247,6 @@ def normalize_skill_name(name: str) -> str:
     """
     # Use to_hyphen_case which already handles most normalization
     return to_hyphen_case(name)
-
-
-# =============================================================================
-# Package Type Routing Functions (T4)
-# =============================================================================
-# These functions determine behavior based on:
-# 1. Explicit `type` field in apm.yml (highest priority)
-# 2. Presence of SKILL.md at package root (makes it a skill)
-# 3. Default to INSTRUCTIONS for instruction-only packages
-#
-# Per skill-strategy.md Decision 2: "Skills are explicit, not implicit"
-# - Packages with SKILL.md OR explicit type: skill/hybrid -> become skills
-# - Packages with only instructions -> compile to AGENTS.md, NOT skills
-
-
-def get_effective_type(package_info) -> "PackageContentType":
-    """Get effective package content type based on package structure.
-
-    Determines type by:
-    1. Package has SKILL.md (PackageType.CLAUDE_SKILL or HYBRID) -> SKILL
-    2. Package is a SKILL_BUNDLE or legacy marketplace plugin with skills/ -> SKILL
-    3. Otherwise -> INSTRUCTIONS (compile to AGENTS.md only)
-
-    Args:
-        package_info: PackageInfo object containing package metadata
-
-    Returns:
-        PackageContentType: The effective type
-    """
-    from apm_cli.models.apm_package import PackageContentType, PackageType
-
-    # Check if package has SKILL.md (via package_type field)
-    # PackageType.CLAUDE_SKILL = has root SKILL.md only
-    # PackageType.HYBRID = has both apm.yml AND root SKILL.md
-    # PackageType.SKILL_BUNDLE = has skills/<name>/SKILL.md (nested bundle)
-    # PackageType.MARKETPLACE_PLUGIN = explicit legacy Claude plugin input.
-    if package_info.package_type in (
-        PackageType.CLAUDE_SKILL,
-        PackageType.HYBRID,
-        PackageType.SKILL_BUNDLE,
-        PackageType.MARKETPLACE_PLUGIN,
-    ):
-        return PackageContentType.SKILL
-
-    # Default to INSTRUCTIONS for packages without SKILL.md
-    return PackageContentType.INSTRUCTIONS
-
-
-def should_install_skill(package_info) -> bool:
-    """Determine if package should be installed as a native skill.
-
-    This controls whether a package gets installed to .github/skills/ (or .claude/skills/).
-
-    Per skill-strategy.md Decision 2 - "Skills are explicit, not implicit":
-
-    Returns True for:
-        - SKILL: Package has SKILL.md or declares type: skill
-        - HYBRID: Package declares type: hybrid in apm.yml
-
-    Returns False for:
-        - INSTRUCTIONS: Compile to AGENTS.md only, no skill created
-        - PROMPTS: Commands/prompts only, no skill created
-        - Packages without SKILL.md and no explicit type field
-
-    Args:
-        package_info: PackageInfo object containing package metadata
-
-    Returns:
-        bool: True if package should be installed as a native skill
-    """
-    from apm_cli.models.apm_package import PackageContentType
-
-    effective_type = get_effective_type(package_info)
-
-    # SKILL and HYBRID should install as skills
-    # INSTRUCTIONS and PROMPTS should NOT install as skills
-    return effective_type in (PackageContentType.SKILL, PackageContentType.HYBRID)
-
-
-def should_compile_instructions(package_info) -> bool:
-    """Determine if package should compile to AGENTS.md/CLAUDE.md.
-
-    This controls whether a package's instructions are included in compiled output.
-
-    Per skill-strategy.md Decision 2:
-
-    Returns True for:
-        - INSTRUCTIONS: Compile to AGENTS.md only (default for packages without SKILL.md)
-        - HYBRID: Package declares type: hybrid in apm.yml
-
-    Returns False for:
-        - SKILL: Install as native skill only, no AGENTS.md compilation
-        - PROMPTS: Commands/prompts only, no instructions compiled
-
-    Args:
-        package_info: PackageInfo object containing package metadata
-
-    Returns:
-        bool: True if package's instructions should be compiled to AGENTS.md/CLAUDE.md
-    """
-    from apm_cli.models.apm_package import PackageContentType
-
-    effective_type = get_effective_type(package_info)
-
-    # INSTRUCTIONS and HYBRID should compile to AGENTS.md
-    # SKILL and PROMPTS should NOT compile to AGENTS.md
-    return effective_type in (PackageContentType.INSTRUCTIONS, PackageContentType.HYBRID)
 
 
 def copy_skill_to_target(
